\name{inputEx}
\docType{data}
\alias{inputEx}
\alias{inputEx-sp}
\alias{inputEx.nona}

\title{A fictitious aCGH data set}
\description{
  A fictitious aCGH data set.
}
\usage{data(inputEx)}

\format{A data frame with 500 rows and 6 columns; the last three
  correspond to the aCGH data for three samples. There are data for five
  chromosomes. The same file is available in three formats: an RData
  file, a tab separated text file, and a space-separate text file.}

\source{Simulated data}
% \references{
%     Diaz-Uriarte, R. and Rueda, O.M. (2006). ADaCGH: an R package and
%   web-based application for the analysis of aCGH data. Tech. report.
%   \url{http://www.ligarto.org/rdiaz/Papers/adacgh.pdf},
%   \url{http://adacgh.bioinfo.cnio.es}.
% }
\keyword{datasets}
