\name{inputEx1}
\docType{data}
\alias{inputEx1}
\title{A fictitious aCGH data set}
\description{
  A fictitious aCGH data set.
}
\usage{inputEx1}
\format{A data frame with 4323 rows and 6 columns; the last three
  correspond to the aCGH data for three samples.}
\source{Simulated data}
% \references{
%     Diaz-Uriarte, R. and Rueda, O.M. (2006). ADaCGH: an R package and
%   web-based application for the analysis of aCGH data. Tech. report.
%   \url{http://www.ligarto.org/rdiaz/Papers/adacgh.pdf},
%   \url{http://adacgh.bioinfo.cnio.es}.
% }
\keyword{datasets}